feat (dev_spring_boot) implemented jwt service interface
