+++ src/main/java/com/project/jwt/JwtServiceImpl.java
@@ -1,0 +1,92 @@
+package com.project.jwt;
+
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+import java.security.Key;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+import java.util.Date;
+import java.util.Map;
+import java.util.function.Function;
+
+@Service
+public class JwtServiceImpl implements JwtService {
+    @Value("${jwt.secret}")
+    private String secretKey;
+
+    @Override
+    public String extractUsername(String token) {
+        return extractClaim(token, Claims::getSubject);
+    }
+
+    @Override
+    public String generateAccessToken(UserDetails userDetails) {
+        return generateToken(Map.of("Type", "Access"), userDetails, 5, ChronoUnit.MINUTES);
+    }
+
+    @Override
+    public String generateRefreshToken(UserDetails userDetails) {
+        return generateToken(Map.of("Type", "Refresh"), userDetails, 2, ChronoUnit.HOURS);
+    }
+
+    @Override
+    public boolean isTokenValid(String token, UserDetails userDetails) {
+        final String email = extractUsername(token);
+        return (email.equals(userDetails.getUsername())) && !isTokenExpired(token);
+    }
+
+    @Override
+    public boolean isRefreshToken(String token) {
+        return "Refresh".equals(extractType(token));
+    }
+
+    private String extractType(String token) {
+        return extractClaim(token, claims -> claims.get("Type", String.class));
+    }
+
+    private <T> T extractClaim(String token, Function<Claims, T> claimsResolver) {
+        final Claims claims = extraxtAllClaims(token);
+        return claimsResolver.apply(claims);
+    }
+
+    private String generateToken(Map<String, Object> extraClaims, UserDetails userDetails, int timeAmount,
+                                 ChronoUnit chronoUnit) {
+        Instant now = Instant.now();
+        return Jwts
+                .builder()
+                .setClaims(extraClaims)
+                .setSubject(userDetails.getUsername())
+                .setIssuedAt(Date.from(now))
+                .setExpiration(Date.from(now.plus(timeAmount, chronoUnit)))
+                .signWith(getSignInKey())
+                .compact();
+    }
+
+    private boolean isTokenExpired(String token) {
+        return extractExpiration(token).before(new Date());
+    }
+
+    private Date extractExpiration(String token) {
+        return extractClaim(token, Claims::getExpiration);
+    }
+
+    private Claims extraxtAllClaims(String token) {
+        return Jwts
+                .parserBuilder()
+                .setSigningKey(getSignInKey())
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    private Key getSignInKey() {
+        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
+        return Keys.hmacShaKeyFor(keyBytes);
+    }
+}
